feat: add year selection on employment page
