+++ src/clients/employments-references.js
@@ -40,8 +40,8 @@ export const findReference = async ({ hospitalId, referencesId, headers }) => {
 }
 
 export const searchReferenceForMonth = async ({ hospitalId, year, month, headers }) => {
-  const query = `?type=searchByMonth&year`
-  const response = await fetch(`${API_URL}${HOSPITALS_ENDPOINT}/${hospitalId}/employments-references/${query}`, {
+  const query = `?type=searchByMonth`
+  const response = await fetch(`${API_URL}${HOSPITALS_ENDPOINT}/${hospitalId}/employments-references${query}`, {
     headers: { ...headers, "Content-Type": "application/json" },
     method: METHOD_POST,
     body: JSON.stringify({ year, month }),
